rename Display Options as Fields
