+++ electron/app/components/FieldsSidebar.tsx
